+++ src/hotaru/filter/gaussian.py
