+++ src/hotaru/footprint/clean.py
@@ -3,7 +3,7 @@ import pandas as pd
 import tensorflow as tf
 
 from ..evaluate.footprint import calc_sim_area
-from ..image.laplace import gaussian_laplace_multi
+from ..filter.laplace import gaussian_laplace_multi
 from ..util.dataset import unmasked
 from ..util.distribute import ReduceOp
 from ..util.distribute import distributed
